Generalized the script
The saving of attachment assumes the attachments has the string date on it, as well as the string "workstation" / "server". This change added a new $attachmentfilter user defined variable for user to define what is the name of the attachment they are looking for specificity. Since the creation of the attachment destination folder is based on the name of the attachment, this approach has been changed to based on the date of the email received instead.